Changed double to float in noise gen to get more interesting farlands, implemented click event for CommandTP

+++ Dependencies/FastNoise/FastNoise.h
@@ -39,7 +39,7 @@
 #ifdef FN_USE_DOUBLES
 typedef double FN_DECIMAL;
 #else
-typedef double FN_DECIMAL;
+typedef float FN_DECIMAL;
 #endif
 
 class FastNoise
